% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_poc_token.R
\name{extract_poc_token}
\alias{extract_poc_token}
\title{Extract token from Oliver sign-in GET response}
\usage{
extract_poc_token(response)
}
\arguments{
\item{response}{A response from a GET to one of the oliver Sign-in Domains containing a "poc.t" cookie.}
}
\description{
Extract token from Oliver sign-in GET response
}
