% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csv_extract.R
\name{get_csv_extract}
\alias{get_csv_extract}
\title{GET Oliver csv extracts via authentication object}
\usage{
get_csv_extract(extract_name = NA)
}
\arguments{
\item{extract_name}{A valid extract name available to the specified user}
}
\description{
GET Oliver csv extracts via authentication object
}
