% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_info.R
\name{get_org_info}
\alias{get_org_info}
\title{GET Oliver organization information}
\usage{
get_org_info(org_ids = NA)
}
\arguments{
\item{org_ids}{A valid organization id or vector of ids. Expects integer vectors.}
}
\description{
This is a function that will hit the Oliver organization API and return information
related to that organization. At this time, the function only returns the organization
name.
}
