% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_auth_to_oliver.R
\name{post_auth_to_oliver}
\alias{post_auth_to_oliver}
\title{Authenticate to the Oliver API}
\usage{
post_auth_to_oliver(oliver_email, oliver_password, oliver_env)
}
\arguments{
\item{oliver_email}{An email registered with the Oliver application.}

\item{oliver_password}{A password associated with \code{oliver_email}.}

\item{oliver_env}{The desired API environment.}
}
\description{
Authenticate to the Oliver API
}
