% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_subdomain.R
\name{return_subdomain}
\alias{return_subdomain}
\title{Define an Oliver subdomain to serve as a base url}
\usage{
return_subdomain(oliver_env)
}
\arguments{
\item{oliver_env}{A character string representing the desired API environment: 'production', 'usertest', or 'staging'.}
}
\description{
Define an Oliver subdomain to serve as a base url
}
